Code style fixes
