+++ src/main/java/org/radarcns/auth/PermissionOnSubjectFilter.java
@@ -9,7 +9,6 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MultivaluedMap;
 import org.radarcns.auth.authorization.Permission;
 import org.radarcns.auth.token.RadarToken;
-import org.radarcns.webapp.filter.AuthenticationFilter.RadarSecurityContext;
 
 /**
  * Check that the token has access to given subject in given project.
