+++ src/main/java/org/radarcns/webapp/GrizzlyServer.java
@@ -1,6 +1,5 @@
 package org.radarcns.webapp;
 
-import java.io.IOException;
 import java.net.URI;
 import org.glassfish.grizzly.http.server.HttpServer;
 import org.glassfish.hk2.api.ServiceLocator;
@@ -9,10 +8,12 @@ import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/** Server for the REST-API. */
 public class GrizzlyServer {
     private static final Logger logger = LoggerFactory.getLogger(GrizzlyServer.class);
     private static final String BASE_URI = "http://0.0.0.0:8080/radar/api/";
 
+    /** Main command line server. */
     public static void main(String[] args) {
         ServiceLocator locator = ServiceLocatorUtilities.createAndPopulateServiceLocator();
 
@@ -22,11 +23,12 @@ public class GrizzlyServer {
         try {
             httpServer.start();
 
-            System.out.println(String.format("Jersey app started on %s.\nHit any key to stop it...", BASE_URI));
+            System.out.println(String.format("Jersey app started on %s.\nHit any key to stop it...",
+                    BASE_URI));
             //noinspection ResultOfMethodCallIgnored
             System.in.read();
-        } catch (IOException e) {
-            logger.error("error starting server: "+e.getLocalizedMessage(), e);
+        } catch (Exception e) {
+            logger.error("Error starting server", e);
         }
     }
 }
