+++ src/main/java/org/radarcns/webapp/filter/AuthorizationFeature.java
@@ -12,6 +12,7 @@ import org.radarcns.auth.PermissionFilter;
 import org.radarcns.auth.PermissionOnProjectFilter;
 import org.radarcns.auth.PermissionOnSubjectFilter;
 
+/** Authorization for different auth tags. */
 @Provider
 public class AuthorizationFeature implements DynamicFeature {
     @Override
