+++ src/main/java/org/radarcns/webapp/filter/CorsFilter.java
@@ -7,6 +7,7 @@ import javax.ws.rs.container.ContainerResponseFilter;
 import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.ext.Provider;
 
+/** Standard CORS filter, applied to all responses. */
 @Provider
 @Priority(1000)
 public class CorsFilter implements ContainerResponseFilter {
