+++ src/main/java/org/radarcns/webapp/validation/AlphanumericValidator.java
@@ -4,6 +4,9 @@ import java.util.regex.Pattern;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
+/**
+ * Checks that a given String should be alphanumeric plus allowed symbols {@code -_./:#@}.
+ */
 public class AlphanumericValidator implements ConstraintValidator<Alphanumeric, String> {
     private static final Pattern USER_PATTERN = Pattern.compile("[a-zA-Z0-9-_./:#@]+");
